Revert DB version 2 to look like that on master
This moves the changes to the _scan function into version 7, so that
they will be properly applied to any running deployments.  It also
renames the `page_offset` argument back to the original `page`, avoiding
changing the calling signature of the functions.
