+++ libraries/entities/test/db/versions/0001.yml
@@ -104,7 +104,7 @@ methods:
     description: test
     mode: read
     serviceName: test-entities
-    args: pk text, rk text, condition text, size integer, page_offset integer
+    args: pk text, rk text, condition text, size integer, page integer
     returns: table (partition_key text, row_key text, value jsonb, version integer, etag uuid)
     # The limit will always be size + 1 to allow consumers to detect when it's the last page.
     # If the list of result > size then there are more entries to fetch. Else, there are no next pages.
@@ -145,8 +145,8 @@ methods:
         if size is not null and size > 0 then
           sql := sql || ' limit ' || size + 1;
 
-          if page_offset is not null and page_offset > 0 then
-            sql := sql || ' offset ' || page_offset;
+          if page is not null and page > 0 then
+            sql := sql || ' offset ' || page;
           end if;
         end if;
